+++ src/main/java/com/github/leandromoraesrj/sincadbackend/domain/RegimeApuracao.java
@@ -23,9 +23,9 @@ public class RegimeApuracao implements Serializable {
 	@SequenceGenerator(name = "se_regime_apuracao_generator", sequenceName = "se_regime_apuracao", allocationSize = 1)
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "se_regime_apuracao_generator")
 	@Column(name = "sq_regime_apuracao")
-	private Long id;	
+	private int id;	
 	@Column(name = "co_regime_apuracao")
-	private Long codigo;	
+	private int codigo;	
 	@Column(name = "no_regime_apuracao")
 	private String nome;	
 	@Temporal(TemporalType.DATE)
@@ -38,7 +38,7 @@ public class RegimeApuracao implements Serializable {
 	public RegimeApuracao() {		
 	}
 
-	public RegimeApuracao(Long id, Long codigo, String nome, Date inicio, Date fim) {
+	public RegimeApuracao(int id, int codigo, String nome, Date inicio, Date fim) {
 		super();
 		this.id = id;
 		this.codigo = codigo;
@@ -47,19 +47,19 @@ public class RegimeApuracao implements Serializable {
 		this.fim = fim;
 	}
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 
-	public Long getCodigo() {
+	public int getCodigo() {
 		return codigo;
 	}
 
-	public void setCodigo(Long codigo) {
+	public void setCodigo(int codigo) {
 		this.codigo = codigo;
 	}
